Thrown out the entire graphics system, visibility and resource. Renamed Frame2 to Frame, replacing the old system.
The new component based graphics system, which is supposed to be more data-oriented than the old system.
Visibility should be reconsidered. There is really nothing wrong with the visibility link creation, but the VisResolver and VisResolveContainer gets reset each time the scene has to be rendered through a view, and the VisResolveContainer::Clear() is high up on the list inclusive samples when doing profiling.