*foundation* Added two new allocators for foundation, one which allocates memory in pools and returns a slice when called, and the other which is *supposed* to allocate a chunk of slices. Added implicit Ptr conversions for types using std::is_base_of with static_assert to ensure the conversion is valid.
*render*
Added class for a camera, which is used to represent the viewing conditions for a View.
Visibility is designed to be done in a lock-step manner, that is, the visibility system will reset its visibility lists to match the scene just before rendering takes place.
Resource tags are now just an array parallel to the rest of the loaded resources, and when resources are removed per tag, they are not grouped in a dictionary, but is instead just in a list.