*foundation* new: - Added a new allocator which uses an ID pool to generate slices of data, which can be used for DOD like entity allocations.
*render*
intro:
The new render system, call it render2, will not be as heavily object-designed as the previous was. Instead, handles represent objects, and they are simply ids within some idallocator. Every subsystem provides a facade type interface to interact with the rendering side, however direct access to the different pools in which graphics objects reside is also possible to access through this facade. Most global graphics servers and resources are transformed to memory pools, and there are two families. Stream pools which stream load files from disk or the net, and memory pools which load directly from the code. This also goes for all previous resources, they are also moved to pools which means there is, for example, no longer a vertex buffer class, but instead there is a VertexBufferId, which is a 32 bit handle for a vertex buffer. The class files remain physically during this transition period, although not in the CMake structure.

And oh, nothing compiles in render atm ;)