[foundation]
Fixed SFINAE issue in HashTable when compiling for C++14 in GCC.
Removed superfluous type alias in array.h
id allocator can now defragment data by calling Defragment and sending the list of currently used IDs. The list will be manipulated (items shuffled around) upon completion. Not sure how useful this is yet, perhaps an array of all deleted indices should be produced as an output?
Fixed enormous bug in idpool which would cause it to malfunction if the number of ids exceeded 512.