[foundation] Added memory mapping interface for Windows (ONLY, BEWARE!). Removed the asserts in ArrayAllocatorSafe, the reason why we have a critical section is to allow us to enter it on several threads...
[render]
Changed the FenceWait to FenceWaitAndReset, which is more clear. Also added an ordinary fence wait which doesn't reset.
Added a handover submission context, which is meant to handover resources from transfer to graphics, which can be triggered from another thread.
Using the new memory mapping in both texture and mesh streaming.
Increased the index type in imgui to 32 bit, maybe stupid, we'll see... but I need it for the profiler.
Enabled async loading for textures and meshes, wohoo!
The graphics device now uses a lock for the resource (and handover) submissions, the only reason why there are two contexts is because you need a submission context to act on a single queue.
Added the API for waiting for a semaphore, as supported by the new timeline semaphores. This is used for the handover to not have to wait for the graphics fence, but can instead just wait for the semaphore. Actually, we should use this everywhere where we have fences, as they are no longer required.
Submission contexts can now save which timeline submission index they have for a certain frame index, allowing us to wait for that specific submission on the next frame index.
We now stream in new LODs asynchronously.
Added an API for queuing up new image views to be created in the shader server. This is because we might stream-load new LODs, which require a new image view so that we can see these new LODs.
All model loads now send a callback for when the model is loaded, such that observers and such can actually register their stuff when the model actually loads (which they have to).
Started working on allowing a surface to update the LODs.
Simplified observable setup procedure, it is really stupid that there is an ad-hoc way to setup entities. Instead, when an entity is registered as an Observable, all Observers should react to it on their next update, which is the case now.