Entity system overhaul
Messed up the rebase, so this will be a squash :(

The new entity system is more similar to how the old (N3) system was, but with improved spacial and temporal memory usage.

The entity system now consists of an in-memory, non-relational database that is made to be as fast as possible for insertion, deletion, and iteration.
Entities are bundled together depending on which attibutes they have. All entities that have the same attributes attributed to them will be allocated within one single table, called a category.

Categories are setup via blueprints (json), preferably via some editor, or directly from code.

Attributes can be bundled via properties. This means each category that has a specific property will automatically have all attributes that the property "works" on. Each property can loop over all the data of the category that the property is attached to to perform gameplay updates or similar.

The database is also easily (and efficiently) queryable which means extending the entity system with "systems" (as in ECS system) or "processors" (more fitting name IMO) is fairly straight forward.

Managers has been rewritten to be more aligned with the new design philosophy, meaning they are no longer refcounted object, but instead are just a bundle of functions that are allocated from a pool. This should make it a lot easier to add new types of manager, and also implement managers from dynamically loaded libraries (C hotreloading anyone? :D )

There are still some improvements that can be made.
No IDLC has been implemented yet, though the code that it would generate is actually quite small and should be easy and quick to implement.
No saving or loading the database has been implemented. This should also be quite straight forward.

Messaging is still an open problem. There seems to be no really clever way to fit such a system into the current design, but it might also be just me not seeing the solution.

For now, we should test and evaluate the system and see if it's a good approach or if we should move away from the "property" approach and instead opt for a fully fledged processor approach. Right now, there seems to be use-cases for both.