render: Fixed bugs in terrain and pass
Pass applies the VK_SUBPASS_EXTERNAL sync again. Actually, it's necessary, so keep it this time! It fixed the blinking in the terrain code, which would happen even at low FPS.
Fixed the millions of one-offs in terrain, so now it works, but still doesn't move subtexture mips when subtextures update.
Fixed missing reference use of arrays in graphics device copy.