Terrain changes
Rehauled the extraction process - instead of using a buffer which is cleared in a compute shader, updated when rendering, and extracted, we instead implement a buffer with atomic uints for page residency, which is cleared with a copy every frame, and the whole extraction process is done in the terrain prepass shader.
Cleaned up some dead code.
Subtexture moves - When a subtexture is updated, it's pixels are moved from the old region to the new region by copying the mips which would fit in the new region, then the old area gets cleared to remove any random tiles getting sampled when those pixels are accessed again.