Fixed submission context cleanup.
The Handover and Resource submission contexts cleaned themselves up based on if the semaphore value had passed the submission which the resources were waiting for. This seems to not work, so instead we insert a fence and poll this fence every frame for completion, to know if the resources can be cleaned or not.