+ Managed Properties
Added flags to memdb properties so that we can flags properties with different attributes.

Entity system properties can now be flagged with MANAGED, which essentially delays the destruction of those specific properties, so that managers can cleanup any allocated resources.

GraphicsManager displays an example of this, where it checks for entities which has ModelEntityData, but not ModelResource. All entities that pass this filter will have been deleted and their graphics resources needs to be cleaned up.