Added queue support flags for buffer
Previously, based on the type of buffer, we inferred how the buffer would be shared and used between queues. This has been replaced with a more explicit flag, because it turned out we did a lot of unnecessary cross-queue buffer sharing which only lived on a single queue type. The old behavior still exists, using AutomaticQueueSupport, and is the default when creating a buffer, however these rules ONLY apply on buffers which are transfer sources or destinations. All other buffers require explicit queue declarations. Otherwise, the queue ownership is implicit, i.e. the queue it is used on first will be the queue it belongs to, however it is still strongly advised to set this flag anyways, if not just to make it explicit how we intend to use our buffers.