CommandLineArgs improvements
CommandLineArgs supports finding multiple values assigned to the same flag, for example one can run an application with -i A B C D and all A B C D would happen under the flag -i.
Find and FindIndex now supports finding the next item of an index, which is required to support CommandLineArgs. The new argument defaults however, so the old behavior remains.