Simplified buffer API and added upload for queues
Removed the external call for BufferUpdateArray because an array of elements only makes sense if we have a type, so it remains on the templated versions.
Added methods for uploading data to a buffer while recording a main command buffer, which allows for just-before-draw updates of buffers.
Added call to fetch from the implementation the max size for a command buffer buffer update.