New luminance calculation
Refactored TimeAndRandom.
HistogramContext executes shader which buckets luminance converted color values into a 256 sized array of counters. Each time a value is bucketed, the counter for that bucket increases. The luminance is transformed to a logarithm initially, resulting in an uneven distribution of values having more at low buckets, and fewer in bigger so as to