Memory optimizations
Moved drawpacket to shaderstatenode. Also reduced size of it since we used a fixed size (only modifying dynamic offset group)
ResourceTableLayouts no longer overallocate descriptors unless such hints are given. The overallocationSize hint is passed when creating a resource table to indicate if we are going to need more resource tables of the same type or not. The automatic grow function for this call has been disabled.