Fixed age old barrier sync setup in frame script
Instead of setting up the dependency manually when building, we should always use the textureDeps or bufferDeps in the FrameOp and let the framescript builder resolve it automatically.
Also fixed a bug where barriers of differing destination stages but matching source stages would not become it's own barriers. This problem can be worked around with the new pipeline barriers which allows for the pipeline stages in the individual resource barriers, so consider simplifying this code in the future.