Enables disabling of ops prior to build
FrameScript allows for lookup of ops. When ops are fetched, they can be disabled and the script rebuilt.
Fixed ImGUI so it doesn't crash if the framescript doesn't draw ImGUI.
Generalized FrameScript so that every op can have children. It's useless for a lot of passes and the FrameScriptLoader won't parse it for every op, but it makes it simpler to streamline and generalize child ops.
Added function to search for ops using a pattern, like node1/node2/node3.
Moved the ShaderServer SetupBufferConstants so it doesn't happen implicitly in View. The reason is that we cannot rebuild framescripts if this call is only done when a view is created. Also, removed the ugly CoreGraphics::RegisterTexture pattern. How it should work is this - if the window resizes, all views should call resize with their framescripts.