Long list of changes incoming
Changed all the Pool-systems to be called Cache, as it's more appropriate for what the system is doing.
Removed the node instance stuff from Model. Instead, that stuff is being handled in a much more data-oriented manner by ModelContext.
The Particle and Character specific stuff related to models and rendering is done in ParticleContext and CharacterContext respectively.
Node information, like activity, lod-culling, always visible and if the node has been seen by any observer is now saved as a node flags bitmask.
There is no longer and backpointer from a node instance to the node - instead, the vertex/index buffer, vertex layout and primitive group application is done in a lambda (happens relatively rarely anyways) and the draw call is reduced to being just a CoreGraphics::Draw with a few arguments (instancing to be precise). This allows for a per-object instancing (particles for example) and a per-pass instancing (for example, shadow cascades).