Various changes
Added support for signaling and reseting a job event on the host side.
Refactored JobSyncSignal to JobSyncThreadSignal and added the above mentioned as JobSyncHostSignal.
Added a lock-unlock mechanism for constant updates so one will get assertions when trying to allocate constants before the frame has progressed or after the constants have been consumed.
Added asserting mutex to try to track down the writeInfo access bug that happens sometimes (race condition most likely).
Fixed hang when rendering shadows because the wrong visModelCmd gets picked.
VisibilityContext now waits for the ParticleContext jobs to finish before finishing the visibility sorting - this is to take into account the base instance and num particles when generating draw packets.