Began work on thread safe graphics device
Noticed an issue with how we swap the backbuffers and when we wait for constant updates. We should only swap the backbuffer just prior to rendering the last pass in order to save time waiting for the display (CPU time), and we can safely wait for the previous buffer just as we submitted the frame. This would effectively allow us to update constants everywhere between EndFrame and RenderViews, which opens up for a wider window of constant updates.
Started working on getting the graphics device multi threaded. This would allow us to run framescripts entirely in jobs, which is really useful.