[foundation][render] Moving to new jobs system
Testing out the new pull job system for visibility, as a part of making the modelcontext able to more easily remove entities.
Added a total invocation offset to the function in jobs2.
Added a function for creating a single job instead of dividing by workgroup (basically just a shortcut).
Added move functions to the event functions on all platforms.
Added an is manual check for event to test for systems which require a manual reset event.
Added a type which just inherits from event but sets the manual reset to true, called EventWithManualReset.
Converted brute force visibility system to use jobs2.
Converted ObserverContext to use jobs2.
Added tests for jobs2.