[foundation] Finished new jobs system
The new pull-style job system allows for threads to grab work when they have time, and allows for jobs to setup dependencies with each other using counters (instead of events), although individual dispatches can be waited on if supplied with a system event.

This system has some advantages over the previous one. First, it's easier to setup jobs, using just a context with some pointers to start work. We can split the work in work groups (similar to a GPU) and dispatch work in chunks, or do a single job running on a single thread. Second, the system uses a queue of jobs which threads can pickup. This allows threads to skip over jobs waiting for some other job to finish, and pickup whatever is available. In theory, this should maximize thread utilization more in comparison to the previous system, where jobs where pushed on threads symmetrically and waited on with a single thread-wide barrier. Third, because jobs aren't executed in lockstep, it's completely possible for a job to queue up another job, provided all the context is available and memory won't be deallocated by the first job when it finishes.

**Implementation details**
The job system is a linked list of nodes, which can be traversed and nodes skipped if their dependency hasn't been met, allowing for a thread to pick up another job further down the list. In the future, we should consider adding a priority flag to the dispatch call so jobs can be inserted in some order.

**API details**
The job system allocates a node, an array holding the wait counters and the context for the job in a single alloc. This way, we can also do a single free to clear the memory up, and it simplifies the API, not forcing the user to not have to deallocate the memory allocated for the context and the wait counter array at some later point. Technically, the job system could provide an API for calculating an allocation size in order to keep allocations away from the user, although knowing when a deletion can happen would require the use of an event being signaled at the end of the dispatch.

**Motivation**
In order to better jobify Nebula, the main driver here being the model system, we need a job system that allows for less strict setup and execution of thread work. For the models, we need a job for characters, particles, model transform updates, data fetching, visibility and sorting, and we need them all to run in very specific orders with very specific dependencies so as to maximize performance. This wasn't possible before with the input/output slices approach where every job has a very specific data flow.