[foundation] Jobs2 appends jobs at end
Jobs were appended to the beginning of the chain, not the end. This way makes more sense.
Also, a job sequence doesn't lock the job thread while it's looping through the jobs, but connects all the sequence nodes and when done, attaches the sequence to the end of the list. Should be slightly more efficient.
Fixed bug with job alloc running out of memory.