[tests] Minor testviewer changes
Made it so every thread in the profiler has some space dedicated to it. This makes the UI jump around less when a thread doesn't have any work to do in a frame.
Playing around with the number of threads to use for the new jobs system. It should be some factor of the number of available cores though, just not sure how many yet.
Added definition to silence iterator base class warnings from C++17.