[render]
Changed the WaitForAllQueues to also clear up any pending deletes and wait for the device to be idle. It's now called WaitAndClearPendingCommands.
Moved the GraphicsDevice pending deletes queue to instead belong in the context. It makes much more sense because a buffer/texture should be deleted when the submission it was last in has finished, and that's what submissions are for.
Added delayed delete image view to submission context.
Added cleanup pending deletes to SubmissionContext. Use with care though, since it's only safe to clean up the submission context when the device has been waited for.