[render]
Window resize is called on graphics contexts too now, and are called from the GLFWGraphicsDisplayEventHandler instead of directly from GLFW window.
HistogramContext and LightContext both depend on the window buffer, and they used to update the resource table every frame to account for any possible change in the frame script (resizing). Now we actually implement the OnWindowResize callback in GraphicsContext to react to this event.