Feature/resize (#55)
* Initial commit for continued work on resize

Simplified pass creation process slightly by putting the subpass related stuff in it's own function.
Added method for TextureView to be recreated if the underlying texture has been changed. This is because the backend creates a view for a certain texture,

* Camera has now GetView and GetTransform

GetView is the inversed transform, and transform is the local->world space transform for the camera.

* [render]

Working on resize.
Fixed validation error complaining about incorrect subpasses.

* [render]

Cleaning up some texture stuff and some fixes related to pass reload.

* [render]

Fixed LightContext for new CameraContext API.

* [render]

Running WindowResize now actually resizes the window.

* [render]

Added a guard to avoid deleting and creating a new ResourceTable and ContantBuffer when a pass is resized, and instead only updates the buffer.
Removed the VkDevice access for the different SetX in ResourceTable.
Deleting a resource table if a write is on the PendingTableCommits queue will delete it from the queue so it won't be written to after deletion. Optionally, we could put the pending delete on a queue too, if that proves to be more robust.

* [render]

Fixed some forgotten renaming updates.

* [render]

FrameScript rebuild now properly clears the events and barrier arrays.
FrameSubmission also resets the resource reset barriers, to properly update the textures used by the barriers.

* [render]

Fixed missing callback detach in GLFW window.

* [render]

Fixed barriers incorrectly resetting when allocating a new object.

* [render]

Minor refactors in Frame.
Fixed compilation issue in VkPass.

* [render]

Textures only have their bindless handle deregistered if it's not 0xFFFFFFFF.
Reloading circumvents unregistering the bindless texture ID by temporarily setting the bindless handle to 0xFFFFFFFF and then resetting it after unload. This allows us to use the Reregister to just bind a new texture to the bindless structure without having to change the bindless ID everywhere.
Also, window texture should also reload, that was a bug.

* [render]

FrameSubpassFullscreenEffect reloads bound textures when window is resized. Most commonly we would use bindless so this isn't necessary to consider, but we might also add this since we do support binding texture values directly.

* [render]

Fixed shadow rendering.

* [render]

Changed the WaitForAllQueues to also clear up any pending deletes and wait for the device to be idle. It's now called WaitAndClearPendingCommands.
Moved the GraphicsDevice pending deletes queue to instead belong in the context. It makes much more sense because a buffer/texture should be deleted when the submission it was last in has finished, and that's what submissions are for.
Added delayed delete image view to submission context.
Added cleanup pending deletes to SubmissionContext. Use with care though, since it's only safe to clean up the submission context when the device has been waited for.

* [render]

Window resize is called on graphics contexts too now, and are called from the GLFWGraphicsDisplayEventHandler instead of directly from GLFW window.
HistogramContext and LightContext both depend on the window buffer, and they used to update the resource table every frame to account for any possible change in the frame script (resizing). Now we actually implement the OnWindowResize callback in GraphicsContext to react to this event.

* [tests]

Fixed incorrect camera setup with the SetTransform -> SetView change.

* Window resize only happens once per frame

We suppress the results from vkAcquireImage and vkQueuePresent to handle suboptimal and out of date (both signs that the window buffer is out of date.

* Fix clang compile issue.

* Fixed bad aspect ratio when window resizes.

* ClusterContext deals with resize

* SSAOContext respects resize

* Fixed clang compilation error

* Fixed Jobs2 shutdown

* Added Set for safe allocators

* Reworked delayed delete to be more generic

* Deleted old jobs system

* Removed deprecated resource deletion

* Fixed resource loader thread not stopping

* Missing namespace fixed

* Fixed bloom resize and incorrect dispatch size

* Filestream fix

Fixed filestream unmapping a memory mapped pointer if closed before unmapping.

* Cleanup Jobs2 in testviewer

* Make sure to clean up map handle