Deprecation of graphics device
The graphics device is no longer the main point for rendering operations. Instead, rendering operations go directly to a command buffer. This allows us to record command buffers in a parallel way, and it serves to clean up a lot of ugly hacks in the graphics device. It also generalizes a lot of code so we can move a lot of logic out from the graphics device and do it in the higher level API instead.
Submission context is gone. Instead, this is handled in a much more light-handed way in the graphics device.