Command buffer integration
We removed the different constant buffer allocations since the allocation function was already atomic, so all places where we allocate have been changed to reflect that.
All contexts rendering stuff is now doing that to a command buffer, according to the new API.
All frame plugins have been replaced with frame subgraphs, which is the better approach.