Feature/command buffer api (#57)
* Deleted rendezvous

Deprecated unused synch primitive.

* Deprecation of graphics device

The graphics device is no longer the main point for rendering operations. Instead, rendering operations go directly to a command buffer. This allows us to record command buffers in a parallel way, and it serves to clean up a lot of ugly hacks in the graphics device. It also generalizes a lot of code so we can move a lot of logic out from the graphics device and do it in the higher level API instead.
Submission context is gone. Instead, this is handled in a much more light-handed way in the graphics device.

* Frame refactor

Frame uses the new command buffer API.
Removed FrameGlobalState as it's not used.
Added a new frame script node, FrameCode, which can be used to run code and setup dependencies in the code side.
Added a new frame script node, FrameSubgraph, which can group several code-implement frame script ops under a name, which the frame script can inject on build.

* View minor refactor

The view has to build the frame script a bit later, after all the subgraphs have been registered.

* Command buffer integration

We removed the different constant buffer allocations since the allocation function was already atomic, so all places where we allocate have been changed to reflect that.
All contexts rendering stuff is now doing that to a command buffer, according to the new API.
All frame plugins have been replaced with frame subgraphs, which is the better approach.

* Added IDs to natvis, thank me later

* Added enum comparisons with signed int

* Deleted deprecated sync stuff.

* Fixed asserting mutex

It was working atomically when obviously that's what it SHOULDN'T do. This tool is invaluable to find race conditions.

* Added a read write lock mechanism

Replaced the critical section in the safe array allocator
Added a Win32 read write lock implementation.
Change the EnterGet/LeaveGet to Lock/Unlock with locking flags.

* ImGUI color space

ImGUI has to output color values inverting the sRGB conversion so the color comes out correctly as it's put in.

* Linux fixes
Added posix read write lock
Fixed various compilation issues

* Spinlock

Implemented a spinlock mechanism, which is a light-weight, but potentially dangerous locking mechanism, which also supports moving, unlike std mutexes.

* Resource table uses spinlock

It's technically not possible to move a critical section or std::mutex, so they need to use a spinlock mechanism, which incidentally might also be much more light-weight.

* Fixes

* Spinlock

Changed from long to int.

* Fix

* Linux fixes

* Updated for new API