New vertex allocation method
Vertices are allocated from a single global cache, allowing for accessing all vertex data directly in a shader as a single buffer.
This method also circumvents the need to allocate a buffer for vertices. The only downside is that the amount of vertex memory is limited, and has to be setup manually in the CreateGraphicsDeviceInfo.