Bug/preallocated constant memory atomic (#59)
* Minor changes to range allocator.

* New vertex allocation method

Vertices are allocated from a single global cache, allowing for accessing all vertex data directly in a shader as a single buffer.
This method also circumvents the need to allocate a buffer for vertices. The only downside is that the amount of vertex memory is limited, and has to be setup manually in the CreateGraphicsDeviceInfo.

* Small refactor, but this was actually wrong too...

* Revert "Small refactor, but this was actually wrong too..."

This reverts commit 8f7d26d7fcecf058f0e4ab6e0cb81cbf0c241207.

* Revert "New vertex allocation method"

This reverts commit 253f60d3ee59b7280e9d3b8f8e05a142a8bb9a13.

* Revert "Minor changes to range allocator."

This reverts commit 45d959f1c4b9905c70b01b6982336f4810ec64db.

* Fixed bug

Preallocated constant memory wasn't allocated atomically, causing blinking in rare cases.