Feature/global vertex cache (#58)
* Minor changes to range allocator.

* New vertex allocation method

Vertices are allocated from a single global cache, allowing for accessing all vertex data directly in a shader as a single buffer.
This method also circumvents the need to allocate a buffer for vertices. The only downside is that the amount of vertex memory is limited, and has to be setup manually in the CreateGraphicsDeviceInfo.

* Small refactor, but this was actually wrong too...

* Revert "Small refactor, but this was actually wrong too..."

This reverts commit 8f7d26d7fcecf058f0e4ab6e0cb81cbf0c241207.

* Revert "New vertex allocation method"

This reverts commit 253f60d3ee59b7280e9d3b8f8e05a142a8bb9a13.

* Revert "Minor changes to range allocator."

This reverts commit 45d959f1c4b9905c70b01b6982336f4810ec64db.

* Small optimization to range allocator

* Added global vertex cache

The global vertex cache allows for all vertex data to be accessed from shaders. It also prevents buffer allocations for meshes. The downside is that the vertex memory size is decided at device creation, so if more memory is needed, that buffer has to be grown.

* Mesh loader and nodes use new vertex system

Primitive and character nodes utilize the new vertex allocation system.
CharacterSkinNode simplified somewhat since it's just a glorified primitive node. In essence though, there should be one character skin node per 255 joints (for vertex compression reasons), instead of having the fragments stored in the same node. Why? Well because each skin node could have a separate material. But that's a content generation problem.

* Fixed bug with failed resources

Apparently this bug has been here for some time. When loading a resource immediately (not async), the failed resource id was not returned if a second attempt to load the same resource was done.

* Fixed error with Resource Tables

Surprised this wasn't an issue earlier...

* Deprecated MeshBind

We never have access to the Mesh object anyways, and if you do, the actual API is to bind each component individually. No need to hide this away when the only place it's done is inside the model system.

* Minor refactor

The offset to CmdSetIndexBuffer and CmdSetVertexBuffer is actually a byte offset, not an element offset. So it's been renamed as such.

* Minor refactor

* Minor fix

Thread safe array allocator shouldn't use locks for the Get/Set functions, but rather assert a lock is in place.

* Fix

* Small fix