Feature/character constants job (#60)
* Minor changes to range allocator.

* New vertex allocation method

Vertices are allocated from a single global cache, allowing for accessing all vertex data directly in a shader as a single buffer.
This method also circumvents the need to allocate a buffer for vertices. The only downside is that the amount of vertex memory is limited, and has to be setup manually in the CreateGraphicsDeviceInfo.

* Small refactor, but this was actually wrong too...

* Revert "Small refactor, but this was actually wrong too..."

This reverts commit 8f7d26d7fcecf058f0e4ab6e0cb81cbf0c241207.

* Revert "New vertex allocation method"

This reverts commit 253f60d3ee59b7280e9d3b8f8e05a142a8bb9a13.

* Revert "Minor changes to range allocator."

This reverts commit 45d959f1c4b9905c70b01b6982336f4810ec64db.

* Character constants update

Character constants (joints) are updated in a job and this job is waited for much later than before, allowing for a more efficient use of the main thread.

* Characters update async on jobs

* Models update async on jobs

* Fix mistake on character change

* Moved the constant update to job

While we can't really jobify all of the work easily, we could move the constant update to a job, which is done here.