Material changes
More refactoring materialType -> shaderConfig and surface -> material and surfaceInstance -> materialInstance.

Replaced the many mappings of variable to binding per batch and to the default/min/max values to a single name to index mapping, as the per-batch and per-material arrays are identically indexed.

Added constructors for ShaderConfigVariant so you don't need to use the Set and SetType function.

When allocating constants, instead of just returning a uint, we return a ConstantBufferOffset which is just a typedef, but meant to make it more clear that this particular uint can be used as an offset into a constant buffer.

Oh and lastly, added an API for material instance constants. It works by first setting up a MaterialInstanceContext on a ModelContext registered graphics entity. Then, per frame, you run AllocateInstanceConstants to grab memory and setup the resource binding offsets on the model instance level.

Added setting per-instance constants test in the testviewer.