Refactored the way we call graphics contexts
Removed the OnX pattern in favor of 4 arrays of callbacks. This allows contexts to run their functions in any order, no matter how they were created, which gives a benefit in getting an overview in how context functions are called without having to inspect several files.
Consider all entity deletions to be delayed deletes. This seems to be the best one-size-fits-all approach, removing the complexity of the deletion stages and whatnot.
View doesn't just create a new constant struct, but updates it with the per-view settings.